5-20-2013b
Continued work on following the target.  You need to ensure that the
tower tracks the sprite even when it goes behind it, or below the
sprite.  Also, the tower rotates from the upper-left corner not the
center.  So there needs to be some weird translations for the rotation
to look proper.

Perhaps in the interm, move to not having the sprite "track" the enemy,
but just fire projectiles at it in the proper direction.